requirements.txt
absl-py==0.5.0
astor==0.7.1
audioread==2.1.6
autopep8==1.4.3
backcall==0.1.0
bleach==1.5.0
cloudpickle==0.6.1
cycler==0.10.0
dask==0.19.4
decorator==4.3.0
defusedxml==0.5.0
dlib==19.16.0
entrypoints==0.2.3
enum34==1.1.6
future==0.16.0
futures==3.1.1
gast==0.2.0
graphviz==0.9
grpcio==1.15.0
h5py==2.8.0
html5lib==0.9999999
hyperas==0.4
hyperopt==0.1.1
imageio==2.4.1
ipykernel==5.1.0
ipython==7.0.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.1
Jinja2==2.10
joblib==0.12.5
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
keras-rl==0.4.2
kiwisolver==1.0.1
librosa==0.6.2
llvmlite==0.25.0
Markdown==3.0.1
MarkupSafe==1.0
matplotlib==3.0.3
mistune==0.8.4
mock==2.0.0
moviepy==0.2.3.5
nbconvert==5.4.0
nbformat==4.4.0
networkx==2.2
nose==1.3.7
notebook==5.7.0
numba==0.40.1
numpy==1.15.2
opencv-python==3.4.3.18
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pbr==5.1.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
pkg-resources==0.0.0
prometheus-client==0.4.1
prompt-toolkit==2.0.5
protobuf==3.6.1
ptyprocess==0.6.0
pycodestyle==2.5.0
pydot==1.2.4
Pygments==2.2.0
pymongo==3.7.2
pyparsing==2.2.2
pyserial==3.4
python-dateutil==2.7.3
python-speech-features==0.6
pytz==2018.5
PyWavelets==1.0.1
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.1
resampy==0.2.1
scikit-image==0.14.1
scikit-learn==0.20.0
scipy==1.1.0
seaborn==0.9.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
style==1.1.0
tensorboard==1.12.2
tensorflow==1.11.0
tensorflow-estimator==1.13.0
tensorflow-gpu==1.12.0
tensorflow-tensorboard==0.4.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
Theano==1.0.4
toolz==0.9.0
torch==0.4.1
torchvision==0.2.1
tornado==5.1.1
tqdm==4.26.0
traitlets==4.3.2
update==0.0.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2